Everything needed in code for a functional tower
FOR UNITY

HOW TO USE:

Add this script to your turret game object.

The tower mesh needs to be divided in two parts: the stationary part and the moving part.
The moving part is the one you assign as the "top" part in the inspector.

You must also make an empty game object as a child of the turret, this game object needs a sphere collider (with the isTrigger checkbox checked). You then assign this as the "view Distance Sphere" in the inspector of the turret, and also add the TowerViewSphere.cs script. (this is what the turret uses to aim at and shoot enemies marked with the tag "enemy").

You now have a working turret!